+++ src/LIN3S/SharedKernel/Infrastructure/Symfony/Bundle/DependencyInjection/Lin3sSharedKernelExtension.php
@@ -1,0 +1,29 @@
+<?php
+
+/*
+ * This file is part of the Shared Kernel library.
+ *
+ * Copyright (c) 2016-present LIN3S <info@lin3s.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace LIN3S\SharedKernel\Infrastructure\Symfony\Bundle\DependencyInjection;
+
+use Symfony\Component\Config\FileLocator;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Loader;
+use Symfony\Component\HttpKernel\DependencyInjection\Extension;
+
+/**
+ * @author Beñat Espiña <benatespina@gmail.com>
+ */
+final class Lin3sSharedKernelExtension extends Extension
+{
+    public function load(array $configs, ContainerBuilder $container)
+    {
+        $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
+        $loader->load('services.yml');
+    }
+}