+++ src/LIN3S/SharedKernel/Infrastructure/Symfony/Bundle/Lin3sSharedKernelBundle.php
@@ -11,9 +11,6 @@
 
 namespace LIN3S\SharedKernel\Infrastructure\Symfony\Bundle;
 
-use LIN3S\SharedKernel\Infrastructure\Persistence\Doctrine\ORM\Types\PhoneType;
-use LIN3S\SharedKernel\Infrastructure\Symfony\Bundle\DependencyInjection\Compiler\DoctrineORMCustomTypesPass;
-use Symfony\Component\DependencyInjection\Compiler\PassConfig;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -24,34 +21,26 @@ class Lin3sSharedKernelBundle extends Bundle
 {
     public function build(ContainerBuilder $container)
     {
-        $container->addCompilerPass(new DoctrineORMCustomTypesPass(), PassConfig::TYPE_OPTIMIZE);
-
         $container->loadFromExtension('doctrine', [
             'orm' => [
                 'mappings' => [
-                    'Lin3sSharedKernelEmail'   => [
+                    'Lin3sSharedKernelEmail' => [
                         'type'      => 'xml',
                         'is_bundle' => false,
                         'dir'       => $this->basePath() . '/Email/Mapping/',
                         'prefix'    => 'LIN3S\SharedKernel\Domain\Model\Email',
                     ],
-                    'Lin3sSharedKernelSlug'    => [
+                    'Lin3sSharedKernelSlug'  => [
                         'type'      => 'xml',
                         'is_bundle' => false,
                         'dir'       => $this->basePath() . '/Slug/Mapping/',
                         'prefix'    => 'LIN3S\SharedKernel\Domain\Model\Slug',
                     ],
-                    'Lin3sSharedKernelLocale'  => [
+                    'Lin3sSharedKernelPhone' => [
                         'type'      => 'xml',
                         'is_bundle' => false,
-                        'dir'       => $this->basePath() . '/Locale/Mapping/',
-                        'prefix'    => 'LIN3S\SharedKernel\Domain\Model\Locale',
-                    ],
-                    'Lin3sSharedKernelZipCode' => [
-                        'type'      => 'xml',
-                        'is_bundle' => false,
-                        'dir'       => $this->basePath() . '/ZipCode/Mapping/',
-                        'prefix'    => 'LIN3S\SharedKernel\Domain\Model\ZipCode',
+                        'dir'       => $this->basePath() . '/Phone/Mapping/',
+                        'prefix'    => 'LIN3S\SharedKernel\Domain\Model\Phone',
                     ],
                 ],
             ],